requirements.txt
flask==3.1.0
flask-cors==5.0.0
pandas==2.2.3
numpy==2.2.3
scikit-learn==1.6.1
python-dotenv==1.1.0
openpyxl==3.1.5
gunicorn==23.0.0
google-genai==1.68.0
